fix(rsync): Ignore exit code 24 "some files vanished"
As said in the code comment, exit code 24 is used by rsync when a file is
deleted in the middle of the rsync.
But that exit code is not a real exit code:

- rsync does its work like if there is no error
- if there is another, more important, error, rsync prioritize it and uses its
  exit code.

So, it's not satisfying but it's fine to ignore the rsync exit code 24.
[The upstream official answer] suggests exactly that.
[The upstream official answer] also remove the warning log. On my side I think
it's cool to keep it, for example for debugging.

To verify this commit I use `fallocate -l 1G z_big_DB` and
`fallocate -l 1k delete_me` commands to create fake files in a small CouchDB
instance and I checked that, during a backup, the rsync warning was logged, the
exit code ignored, and the `tar.gz` file created contains the expected CouchDB
shards files.

[The upstream official answer]: https://bugzilla.samba.org/show_bug.cgi?id=10356