backup: Don't copy CouchDB data twice on disk
This commit is a BREAKING CHANGE on the `backup` action that drops
options `--rsync-reuse-dir` and `--tmp-dir` in exchange for enhanced
archive creation speed and disk I/O usage.

*Problem*

Currently, all source CouchDB data is transfered over the network and
stored on target disk (with `rsync`). Once this is done (not before), an
archive is created and compressed (with `tar` and `pigz`). This results
in a lot of data being written twice, including the intermediate step
when data is uncompressed (in my use case, this data is ~75 GB, while
the result backup file is only ~25 GB).

*Solution*

This commit proposes another approach: perform the 2 steps in parallel,
and avoid the intermediate disk write (the big uncompressed one). To
achieve this, the source machine creates the `tar` archive and steams it
over the network, while the target continuously reads and compresses it
(with `pigz`) directly to the destination file.

Performing a `tar` instead of a `rsync` shouldn't have any noticeable
CPU impact on source machine, as both commands just "read and send". On
target machine we can expect a higher CPU and resource load as data is
received and compressed at the same time, but this comes with
advantages: less disk I/O, less disk space needed, faster execution.

*Results*

In my use case, the amount of data written to target disk dropped from
100 GB to 25 GB, and the two sequential steps that each took ~17 min to
execute are now done in parallel.