load: Enable CORS
When loading an archive to access a CouchDB instance on `localhost`, it
can be useful to access the database from a redirect or reverse proxy
like nginx. In such a case, the origin might be `localhost:4200` instead
of `localhost:60651`, for example.

To allow this to work in a browser, I propose to enable CORS by default
on `couchcopy load …`.